Increase the default test `duration` a bit
There's a bunch of tests which are executed under the `virtual`
provision plan which can take a little bit more time. I suggest to
increase the default `duration` as I'm not sure it's really worth
to fine tune the value individually and `10m` sounds reasonable.
I believe we don't have to be that strict.