Re-use "become" image to fix sudo-related trouble in /tests/provision/user (#2892)
The test started to fail because `/usr/bin/flock` is not preinstalled
and the default Fedora image does not come with password-less `sudo`
enabled out of the box. Therefore we use the image we build for
`/tests/provision/become`, give a nice name other images will follow
Soon(TM), and use it in a test that exercises non-privileged access.