Improve rendering of story links in our docs (#3001)
There were several broken links, caused by links being misrendered. I
added several comments and removed unused parameters in the process, but
the important bit is the chaneg to rendering of plan/test links, usualy
pointed to by `verified-by` links. No longer adding suffixes or
composing URLs, we rely on `web_link` completely, and if a link points
to multiple objects, like a hierarchy of multiple tests with the same
prefix, we render all of them. It turns out to be hard to get web link
correctly for such a root node, and it provides nice list of all related
links to the user.

Co-authored-by: Petr Matyas <p.matyas13@gmail.com>
Co-authored-by: Petr Šplíchal <psplicha@redhat.com>