Save results before raising exception (#2996)
This is essential for storing at least partial results and making `report`
work if there was some issue with the system (e.g. became unreachable)
during the execution. Let's fetch all artifacts stored in the plan data directory
after each test execution to prevent losing useful information if the
guest becomes unresponsive.

As part of the solution, when waking up an unfinished step, we should not
clear the workdir unless the step in question is actually enabled, which covers
the scenario when user wants to give the step another try.

Co-authored-by: Petr Šplíchal <psplicha@redhat.com>