base: Honor `.gitignore` for rsync (#3037)
I am guessing I may be one of the first people to try using tmt
with a Rust project. The default for the `cargo` toolchain
is to keep a *lot* of cached incremental data in `target/`.
In my case with bootc, it's currently 20G.

A plain rsync() of this is *incredibly* inefficient. rsync doesn't
even use reflinks if available, though that's a distinct bug.

Co-authored-by: Colin Walters <walters@verbum.org>
Co-authored-by: Miloš Prchlík <mprchlik@redhat.com>